Remove display licenses from run/3.
For display licenses, run_display_licenses function should be called.
Split of from run, because display licenses processor exits with process::exit,
which is more difficult to track especially during tests (test binary exited).
Process::exit should be replaced by an exit signal through Result, or Option, or
another combinator, but that is out of scope for this branch.